The band is drawn the way btop draws a graph (#232)
Read `Draw::Graph` in btop's own source rather than recalling it, at the
reader's direction, after two rounds of the masthead still reading as
blocks. Three things it does that this band did not.

One value per sub-column. A 2x4 cell carries two, older on the left,
which is why btop's graphs read as a line where a stretched value reads
as a block. The band now draws one value per sub-column and the glyph
ladder packs the pair, so braille doubles the horizontal resolution
instead of being refused.

An axis. btop's network graph is the closest analogue to this signal,
bursty and zero most of the time, and it is built with `no_zero`, which
forces one dot on the bottom row. A quiet stretch draws a floor rather
than nothing, and a narrow spike then stands on something instead of
floating.

A scale that is not the maximum. `graph_max` is 1.3 times a recent mean
and anything above it saturates. One `cargo build` rewriting a lock file
is two orders of magnitude above an ordinary save, and against a peak
denominator every edit for the next two minutes draws one level high: a
graph that goes blank because something happened.

This supersedes #223's coarsening. That row diagnosed the right defect
and reached for the wrong fix: the answer to a hairline between blanks is
the floor, not a wider column, and with the floor drawn the coarsening
only cost resolution. #158 is narrowed rather than reversed, because a
window holding nothing at all still draws nothing, which is the case it
actually came from.

`level_of` and `band_cell` are gone: both elements now scale through
`level_to` and spell the result through `Glyphs::glyph`, which is one
drawer where there were two.

+++ crates/vigia/src/view.rs
@@ -58,7 +58,7 @@ pub struct FileEntry {
     ///
     /// All zeroes for a file `vigia` has not seen change, which is the
     /// ordinary case for a worktree that was already dirty at startup.
-    pub spark: [u16; HISTORY_BUCKETS],
+    pub spark: [u32; HISTORY_BUCKETS],
     /// How recently this file changed, which is what dims a settled row and
     /// what puts the pulse on one that just moved.
     pub recency: Recency,
@@ -476,7 +476,7 @@ pub struct View {
     /// bucket draws a track, so a peak of zero means every bucket is empty and
     /// every one of them is still drawn. `vigia_core::History::peak` carries the
     /// same correction, and this is the copy a renderer actually reads.
-    pub peak: u16,
+    pub peak: u32,
     /// The whole worktree's churn over the window, oldest sample first.
     ///
     /// **What the masthead's band draws** ([#158](https://github.com/breferrari/vigia/issues/158)),
